% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/means_variances.R
\name{ceq1}
\alias{ceq1}
\title{Equilibrium expected value -- single cell type.}
\usage{
ceq1(lambda, beta, delta)
}
\arguments{
\item{lambda}{autonomous rate of cell formation per unit of time}

\item{beta}{proportion of cells that replicate per unit of time}

\item{delta}{proportion of cells that die per unit of time}
}
\value{
the equilibrium expected cell count
}
\description{
Equilibrium expected value -- single cell type.
}
\examples{
ceq1(10, .01, .02)
}

