% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/means_variances.R
\name{cond_ev}
\alias{cond_ev}
\title{Conditional expectation of next step of process}
\usage{
cond_ev(Y, lambda = parms$lambda, birth = parms$birth,
  death = parms$death, tran = parms$tran, parms, verbose = FALSE)
}
\arguments{
\item{Y}{vector of cell counts at time t}

\item{lambda}{autonomous rate of cell formation for first cell type per unit of time}

\item{birth}{vector for each type of cell with proportions of cells that replicate per unit of time}

\item{death}{vector for each type of cell with proportions of cells that die per unit of time}

\item{tran}{vector of length equal to length(Y) -1 with proportions of cells that transform to the next cell type per unit of time}

\item{parms}{option list with parameters lambda, birth, death and tran}

\item{verbose}{(default FALSE) print additional output for debugging}
}
\value{
conditional expectation at time t+1
}
\description{
Conditional expectation of next step of process
}
\examples{
cond_ev(c(100,100,100), 10, .01, .03, c(.01,.02))
parms <- list(lambda = 10, birth = rep(.01,3),
              death = .03, tran = c(.01,.02))
cond_ev(c(100,100,100), parms = parms)

Nsteps <- 2000
mat <- matrix(NA, 3, Nsteps)
mat[,1] <- c(100,100,100)
for ( i in 2:Nsteps) mat[,i] <- cond_ev(mat[,i-1],parms = parms)
# Equilibrium profile with these parameters:
head(t(mat),10)
tail(t(mat),10)
}

