% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/means_variances.R
\name{cond_k}
\alias{cond_k}
\title{Conditional expectation and variance in k steps}
\usage{
cond_k(k, Y, lambda = parms$lambda, birth = parms$birth,
  death = parms$death, tran = parms$tran, parms, verbose = FALSE)
}
\arguments{
\item{k}{number of steps into the future for conditional expectation and variance}

\item{Y}{vector of cell counts at time t}

\item{lambda}{autonomous rate of cell formation for first cell type per unit of time}

\item{birth}{vector for each type of cell with proportions of cells that replicate per unit of time}

\item{death}{vector for each type of cell with proportions of cells that die per unit of time}

\item{tran}{vector of length equal to length(Y) -1 with proportions of cells that transform to the next cell type per unit of time}

\item{parms}{option list with parameters lambda, birth, death and tran}

\item{verbose}{(default FALSE) print additional output for debugging}
}
\value{
conditional expectation at time t+k
}
\description{
Conditional expectation and variance in k steps
}

