% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/means_variances.R
\name{cond_var}
\alias{cond_var}
\alias{cond_var_k}
\title{Conditional variance}
\usage{
cond_var(Y, lambda = parms$lambda, birth = parms$birth,
  death = parms$death, tran = parms$tran, parms, verbose = FALSE)

cond_var_k(k, Y, lambda = parms$lambda, birth = parms$birth,
  death = parms$death, tran = parms$tran, parms, verbose = FALSE)
}
\arguments{
\item{Y}{vector of cell counts at time t}

\item{lambda}{autonomous rate of cell formation for first cell type per unit of time}

\item{birth}{vector for each type of cell with proportions of cells that replicate per unit of time}

\item{death}{vector for each type of cell with proportions of cells that die per unit of time}

\item{tran}{vector of length equal to length(Y) -1 with proportions of cells that transform to the next cell type per unit of time}

\item{parms}{option list with parameters lambda, birth, death and tran}

\item{verbose}{(default FALSE) print additional output for debugging}

\item{k}{number of steps for conditional variance}
}
\value{
conditional expectation at time t+1
}
\description{
Conditional variance
}
\section{Functions}{
\itemize{
\item \code{cond_var_k}: Conditional variance in k steps of the process
}}
\examples{
parms <- list(lambda = 10, birth = rep(.01,3),
              death = .03, tran = c(.01,.02))
cond_ev(c(100,100,100), 10, .01, .03, c(.01,.02))
cond_var(c(100,100,100), parms = parms, verbose = T)
}

