% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/means_variances.R
\name{evec_lt}
\alias{evec_lt}
\title{Eigenvector of a lower triangular matrix}
\usage{
evec_lt(A)
}
\arguments{
\item{A}{a lower triangular matrix}
}
\value{
the eigenvectors of A in the order corresponding the
eigenvalues as they apper on the diagonal of A
}
\description{
Eigenvector of a lower triangular matrix
}
\examples{
A <- matrix(rnorm(25), 5,5)
A[col(A) > row(A)] <- 0
A
evec_lt(A) \%*\% diag(diag(A))- A \%*\% evec_lt(A)   # this should be machine 0
}

