% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/means_variances.R
\name{Lexp}
\alias{Lexp}
\title{L matrix as function of parameters}
\usage{
Lexp(birth = parms$birth, death = parms$death, tran = parms$tran, parms)
}
\arguments{
\item{birth}{vector for each type of cell with proportions of cells that replicate per unit of time}

\item{death}{vector for each type of cell with proportions of cells that die per unit of time}

\item{tran}{vector of length equal to length(Y) -1 with proportions of cells that transform to the next cell type per unit of time}

\item{parms}{option list with parameters lambda, birth, death and tran}

\item{verbose}{(default FALSE) print additional output for debugging

The 'L' matrix is a lower triangular matrix that provides the linear
portion of expectation at time $t$ into expectation at time $t+1$}
}
\description{
L matrix as function of parameters
}

