% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/means_variances.R
\name{mean_cond_var}
\alias{mean_cond_var}
\title{Mean conditional variance}
\usage{
mean_cond_var(lambda = parms$lambda, birth = parms$birth,
  death = parms$death, tran = parms$tran, parms, verbose = FALSE)
}
\arguments{
\item{lambda}{autonomous rate of cell formation for first cell type per unit of time}

\item{birth}{vector for each type of cell with proportions of cells that replicate per unit of time}

\item{death}{vector for each type of cell with proportions of cells that die per unit of time}

\item{tran}{vector of length equal to length(Y) -1 with proportions of cells that transform to the next cell type per unit of time}

\item{parms}{option list with parameters lambda, birth, death and tran}

\item{verbose}{(default FALSE) print additional output for debugging}
}
\value{
\eqn{E ( Var( Y_{t+1} | Y_t)) = T E(D) T'} where
\eqn{D} is the diagonal matrix of variances of new cell births,
deaths and transitions and \eqn{T} is the coding matrix of 0s and \eqn{\pm 1}s
to linearly transform the birth, death and transition components into the
cell counts
}
\description{
Mean conditional variance
}
\examples{
parms <- list(lambda = 10, birth = rep(.01,3),
              death = .03, tran = c(.01,.02))
mean_cond_var(parms = parms)
}

