% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/means_variances.R
\name{sim1}
\alias{sim1}
\title{Simulate single cell process}
\usage{
sim1(N, lambda, beta, delta, burnin = 1000)
}
\arguments{
\item{N}{number of observations}

\item{lambda}{autonomous rate of cell formation per unit of time}

\item{beta}{proportion of cells that replicate per unit of time}

\item{delta}{proportion of cells that die per unit of time}

\item{burnin}{starting the process at the equilibrium value, discard this many observations before generating the observation to be retured}
}
\value{
N consecutive observations from the process
}
\description{
Simulate single cell process
}
\examples{
x <- sim1(10000, 10, .01, .02)
mean(x)
sd(x)
plot(x,pch='.')
}

