% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/means_variances.R
\name{sim_seq}
\alias{sim_seq}
\title{Simulation sequence}
\usage{
sim_seq(N, Y = marg_ev(lambda, birth, death, tran), lambda = parms$lambda,
  birth = parms$birth, death = parms$death, tran = parms$tran, parms,
  verbose = FALSE)
}
\arguments{
\item{N}{number of steps to simulate}

\item{Y}{vector of cell counts at time t}

\item{lambda}{autonomous rate of cell formation for first cell type per unit of time}

\item{birth}{vector for each type of cell with proportions of cells that replicate per unit of time}

\item{death}{vector for each type of cell with proportions of cells that die per unit of time}

\item{tran}{vector of length equal to length(Y) -1 with proportions of cells that transform to the next cell type per unit of time}

\item{parms}{option list with parameters lambda, birth, death and tran}

\item{verbose}{(default FALSE) print additional output for debugging}
}
\value{
sequence of randomly generated values at times t+1, ..., t+N
}
\description{
Simulation sequence
}
\examples{
\dontrun{
parms <- list(lambda=10, birth = c(.01,.02,.01), death = .03, tran = c(.02, .04))
matplot(sim_seq(N = 50, parms = parms), type = 'l', lty = c(1,2,3), lwd = 2, col = c('black','red','green'))
legend("right", inset=.05, legend=c("1", "2", "3"), lty = c(1,2,3), lwd = 2, col = c('black','red','green') , horiz=TRUE)
matplot(sim_seq(N = 1000, parms = parms), type = 'l', lty = c(1,2,3), lwd = 2, col = c('black','red','green'))
legend("right", inset=.05, legend=c("1", "2", "3"), lty = c(1,2,3), lwd = 2, col = c('black','red','green') , horiz=TRUE)
}
}

