% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/means_variances.R
\name{sim_step}
\alias{sim_step}
\title{Simulation step}
\usage{
sim_step(Y, lambda = parms$lambda, birth = parms$birth,
  death = parms$death, tran = parms$tran, parms, verbose = FALSE)
}
\arguments{
\item{Y}{vector of cell counts at time t}

\item{lambda}{autonomous rate of cell formation for first cell type per unit of time}

\item{birth}{vector for each type of cell with proportions of cells that replicate per unit of time}

\item{death}{vector for each type of cell with proportions of cells that die per unit of time}

\item{tran}{vector of length equal to length(Y) -1 with proportions of cells that transform to the next cell type per unit of time}

\item{parms}{option list with parameters lambda, birth, death and tran}

\item{verbose}{(default FALSE) print additional output for debugging}
}
\value{
randomly generated value at time t+1
}
\description{
Simulation step
}
\examples{
parms <- list(lambda = 10, birth = rep(.01,3),
              death = .03, tran = c(.01,.02))
sim_step(c(100,100,100), parms = parms)
}

