% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/means_variances.R
\name{var_equation}
\alias{var_equation}
\title{Solution to variance equation using diagonalization of lower triangular matrix}
\usage{
var_equation(A, L)
}
\arguments{
\item{A,L}{matrices for which to find solution to \eqn{V = A +LVL'}}
}
\value{
\eqn{v} where \eqn{V = A +LVL'} if a solution exists
}
\description{
Solution to variance equation using diagonalization of lower triangular matrix
}
\examples{
A <- diag(3) + .1
A
L <- cbind( c(.8,.1,.1), c(0, .7, .1), c(0,0,.6))
L
V <- var_equation(A,L)
V
V - A - L\%*\% V \%*\% t(L)  #' should be machine 0
}

