% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/means_variances.R
\name{var_equation_vec}
\alias{var_equation_vec}
\title{Solution to variance equation using vec}
\usage{
var_equation_vec(A, L)
}
\arguments{
\item{A,L}{matrices for which to find solution to \eqn{V = A +LVL'}}
}
\value{
\eqn{v} where \eqn{V = A +LVL'} if a solution exists
}
\description{
Solves for \eqn{V} in the equation \eqn{V = A +LVL'} using the fact that
\eqn{vec(V) = vec(A) + (L \kronecker L) vec(V)}. This may work when
\code{\link{var_equation}} does not because \eqn{L} is not diagonalizable.
}
\examples{
A <- diag(3) + .1
A
L <- cbind( c(.8,.1,.1), c(0, .7, .1), c(0,0,.6))
L
V <- var_equation_vec(A,L)
V
V - t(V)
V - A - L\%*\% V \%*\% t(L)  # should be machine 0
V - var_equation(A,L)
# L that can't be diagonalized
L <- cbind( c(.9,.1,0), c(0, .9, .1), c(0,0,.9))
var_equation(A,L)
var_equation_vec(A,L)
eigen(var_equation_vec(A,L))
}

